Small adjustments
Slightly changed where the gold can spawn so that it's less likely for it to spawn out of bounds on certain screens. Made the border radius of dupes slightly more round to make it a little harder to tell the difference. Also changed powerup and dupe spawning to match the change to gold.