Updated sizing
The player, gold, power-up, and dupe sizing used to be in pixels, but was now changed to be in percentage. This is not only so it matches up with the sizing of other things like obstacles, but also so the sizing scales on different sized screens. Measurement for movement and the grow/shrink power ups have been changed as well to match up with this.